Done editing course name+id
